feat: disable favorites toggle on quays without favorites
